test(gms): keep focused v1 coverage and architecture docs
